fix: MergeExchange hangs at shutdown after abort/cancel (#11718)
Summary:

When tasks have an MergeExchange operator and the tasks gets canceled or aborted, the task becomes a zombie and the process also hangs at shutdown time.

Inside a driver thread,  `in MergeExchange::addMergeSources`, `MergeExchange` creates ExchangeClients that are to be used during `Merge::isBlocked`.   `Task` is aborted/canceled, and Drivers are supposed to be canceled, operators closed, and unnecessary references to the tasks are released. However, for MergeExchange, we do not close the sources (ExchangeClients) at shutdown. As a result, the driver thread is calling ExchangeClient::request, which will then go to the application's ExchangeSource. This is then stuck in a loop since the exchange client is not closed.

Reviewed By: xiaoxmeng

Differential Revision: D66666334