feat(planbuilder): Enable passing compressionKind via tableWrite (#11724)
Summary:

Currently there is no way for users to specify `compressionKind` in the velox plan builder in order for the output to be compressed using one of the supported schemes defined by the `CompressionKind` enum. Instead, we always use the hard-coded `CompressionKind_NONE` value.

This diff fixes this by exposing a new `compressionKind` option in one of the overloaded `tableWrite()` methods, allowing users to specify the compression scheme they want to use when writing the table's output files

Reviewed By: kkaranasos, xiaoxmeng

Differential Revision: D66517547