Refactor: Remove unused-variable in velox/runner/tests/LocalRunnerTest.cpp (#11765)
Summary:

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: kevinwilfong, palmje

Differential Revision: D66834083