Fix initialization order of kSpecialforms
Summary:
Since kSpecialforms is a static, it may be initialized during main before types are actually registered. This can cause failures in the fuzzer. For example, ipaddress types is a custom type that needs to be register before kSpecialforms is actually created.

What we actually want is to use init on use pattern to avoid this issue.

Differential Revision: D66856845