fix: IN does not handle User errors thrown during eval
Summary:
IN does not handle user errors thrown during eval. This means that if the IN is wrapped with a TRY, user errors
will still get thrown rather than producing a null.

The fuzzer tests demonstrated a case of this, Velox allows creating Timestamps that are large enough that they
cannot be represented as milliseconds, there would be an overflow, in this case toMillis throws a user exception.
When evaluating IN, the argument is converted to milliseconds and compared to the set of Timestamps. When
wrapped in a TRY the IN still throws an exception rather than returning NULL.

Differential Revision: D67063377