fix(parquet) : Add support for weak functions to register parquet reader/writers
Summary:
This diff marks the registerParquet*factory functions with the weak attribute. The weak attribute allows the linker to discard the weaker implementation of the registerParquet*factory functions and in place use the stronger one (See: https://en.wikipedia.org/wiki/Weak_symbol).
This should remove the need to compile the same files with different options,  leading to ODRs when both targets are added as dependencies to the same binary.

Differential Revision: D67301778