misc: Add support for CustomVectorFuzzers for custom types
Summary:
We've seen with Presto a number of custom types that are not supported in other engines, and may not 
support the full range of values of their backing physical type.  This makes testing them in Fuzzers challenging
as we cannot enable these types broadly (they're not supported in DuckDB for example) and if we do we can
end up with a lot of spurious errors due to invalid values, making the tests low signal.

To address these issues, I've added an interface CustomVectorFuzzer.  An implementation can be written for
a custom type which tells the VectorFuzzer how to generate a random flat/constant Vector of the particular
value.  Then the ReferenceQueryRunner can register these CustomVectorFuzzers with VectorFuzzer so that it
can generate random valid values and add the type to its list of known scalar types (to be used when
generating random types for example).

I've added an example here for TimestampWithTimeZone and AggregationFuzzer.

This is needed as part of enabling support for custom types not supported in DWRF files in Fuzzers
comparing agains Presto.

Differential Revision: D67344181