fix: Casting Varchar to Timestamp should handle offsets that are not recognized time zones (#11849)
Summary:

When casting a Varchar to Timestamp, Presto Java allows an offset timestamp in place of a time zone.  This 
string is of the form +/-HH:MM:SS.mmm where all units except the hour are optional, the colons are optional, 
the . is optional and may also be a , and this string must be separated by the date/time by a single space.

This is not interpreted as a time zone but rather simply a milliseconds from UTC and is applied in addition to the 
time zone conversion if a session time zone is present. Note that strings that fit this pattern but are valid time
zones e.g. +02:00 are still treated as time zones.

Since this is not a true time zone it is not allowed when casting from Varchar to TimestampWithTimeZone or any
other string to time conversions (at least AFAICT).

This change updates fromTimestampWithTimezoneString to handle this case and return an offsetMillis in place
of a time zone if one is present.  Casting from Varchar to Timestamp has been updated to apply this offset to
the result, while all other locations where this function is called have been updated to throw if offsetMillis is 
present.

Reviewed By: kgpai

Differential Revision: D67182709