Use estimateFlatSize in LocalPartition
Summary:
Using retained size is problematic as it may account for memory shared between
multiple vectors.

For example in the join operators vectors are wrapped in dictionary and shared
between multiple output vectors. It was observed that a Join may produce
vectors that retain over 40MB of data while having a flat size of little over
2MB.

When the size reported is high the LocalPartition operator has to block
frequently reducing query performance.

Differential Revision: D67601403