feat: Add cacheable flag in connector split to support fine-grained per-split cache decision
Summary:
Add cacheable flag to support fine-grained per-split cache decision. As Prestissimo might set
cacheable flag in a split (the split context in the protocol) if the split scheduling follow the soft
affinity.

The followup is to add the support in split reader on velox side and set from presto protocol in Prestissimo worker

Differential Revision: D67678409