Add Duck plan transformation for qualified joins and subqueries
DuckDB has a lateral join (DelimJoin) that it introduces into flattened
subqueries. This removes the lateral join so that we extract a logical
plan that can be used as input to Verax or executed as is.

Adds a way of registering a table to the parser with a type instead of dataset.

This code is strictly a stopgap and will be removed as soon as possible.

TODO: Stop using DuckDB as parser and name resolution.