fix: Avoid running the fuzzer if there are no functions to fuzz (#11933)
Summary:

There are edge cases when the input to the fuzzer results in no signatures to fuzz due to combination of skipped functions and skipped input types. This causes the fuzzer to fail when it tries to generate an expression, specifically failing in `ExpressionFuzzer::fuzzReturnType()`, at VELOX_CHECK(!signatures_.empty(), "No function signature available.");. 

Instead of failing the fuzzer at fuzzing time, let's exit the fuzzer with a warning that there were no functions to even fuzz.

Reviewed By: xiaoxmeng

Differential Revision: D67153128