fix: reduce range of buckets in ClassificationAggregationInputGenerator
Summary:
In the AggregationFuzzer we see plans that use the classification_precision aggregation taking upwards of 2
minutes to execute.  The problem seems to be in sorting the results for the purposes of comparison.  The
aggregate produces arrays with size number of buckets and the cost of comparing large arrays of variants is
quite high.

Given we execute 40+ variations of a plan this causes the AggregationFuzzer to run for over an hour on a single
query using this aggregate.

As a quick fix, this change lowers the maximum value of buckets from 50,000 to 500. With this change
executing a single plan takes closer to 10 seconds.

We should probably look at further optimizations in the future, e.g. executing plans in parallel, evaluating sorting 
the Vectors instead of the variants, use a hash based approach.

Differential Revision: D67872087