fix: JsonStringUtil attempts to serialize bad char* in error message
Summary:
JsonStringUtil uses VELOX_USER_CHECK_NE/VELOX_USER_CHECK_LE to compare char* to make sure one is 
not at or beyond the end of the string.  The way the macro works, if it fails, it will attempt to serialize the char*
in the error message.  It doesn't not serialize the address but rather a string starting from the pointer (since it's
a char*).  Given this check has already determined the pointer is at or beyond the end of the string this
produces results that are useless at best.

Using VELOX_USER_CHECK avoids this serialization.

Caught via ASAN with expression fuzzer tests.

Differential Revision: D67875678