feat: Add fast path for single bit in ByteOutputStream::appendBits
Summary:
While working on speeding up the PrestoBatchVectorSerializer I noticed that sometimes we write a single bit to
ByteOutputStreams (e.g. when writing a single null or the null flag).  ByteOutputStream::appendBool has a fast path for this
case, which ByteOutputStream::appendBits is missing, this provides a minor speed up that adds up over time.

Differential Revision: D67988365