fix: Reuse vector in LocalPartition (#12002)
Summary:

X-link: https://github.com/facebookincubator/nimble/pull/122

More than 10% of the CPU are spent on the destruction of local partition output when the load is high.

Also add some optimizations for serialization.  Optimization on `ByteOutputStream::appendBool` does not show significant gain in the query in example (because they are a lot small batches), but it is net gain and would be significant in large batches, so I leave it in the code.

Differential Revision: D67742489