feat(fuzzer): Support multiple joins in the join node "toSql" methods for reference query runners (#11801)
Summary:

Currently, the hash join and nested loop join "toSql" methods for all reference query runners only support a single join. This change extends it to support multiple joins, only needing the join node of the last join in the tree. It traverses up the tree and recursively builds the sql query.

Reviewed By: kevinwilfong

Differential Revision: D66977480