test: Reenable set_agg and set_union in AggregationFuzzer
Summary:
set_agg and set_union were disabled in AggregationFuzzer because Velox and Presto 0.288 had
different ordering semantics.  This was fixed in 0.289 and we bumped the version of Presto we
compare against to 0.290 so this is no longer a concern.

Ran a biased AggregationFuzzer run against only set_agg and set_union comparing against 
Presto for 1 hour and didn't see issues.

Differential Revision: D68286084