test: Add a case with LocalPartition using hash partitioning in JoinFuzzer
Summary:
Today JoinFuzzer only tests plans with LocalPartition that uses round robin partitioning. This
sends the entire Vector to a single thread in addInput, which means we aren't testing the case
where LocalPartition wraps the input in a Dictionary.

I discovered this while investigating the issue fixed in 
https://github.com/facebookincubator/velox/pull/12090  This case, combined with ASAN, is 
enough to catch that issue. (I was able to demonstrate this by applying this change to the 
commit just before the fix).

Differential Revision: D68342589