misc(build): Add variable to configure dependency build parallelism
This change adds the BUILD_THREADS environment variable. This controls the
parallelism used when building dependencies. By default, the number of cores
in the host machine is used but can result in OOM kills. For example,
on machines with 8 cores and 16GB ram (plus 16GB swap) OOM kills can be
observed. This change allows a user to configure a flexible build automation.