misc(fuzzer): Support custom result verifier with compare() API in window fuzzer (#12148)
Summary:

The custom result verifiers of some functions provides a compare() API that can be used to 
compare the expected and actual results after certain transformation. This is needed for 
functions like `avg(interval) -> interval` because the result of this function can have a small 
precision error that should be tolerated. However, the window fuzzer used to only support 
the verify() API of custom result verifiers. This diff makes the window fuzzer to use the 
compare() API as well when available.

Reviewed By: kgpai

Differential Revision: D68507422