Add capacity() getter to ApproxMostFrequentStreamSummary (#12212)
Summary:

# Context

This diff adds a simple getter to retrieve the capacity parameter (after we hit setCapacity()).

We use this internally to perform some sanity checking when we have potentially two different SteamSummaries and we want to ensure capacities are in a sane spot.

This seems reasonably safe to add, the capacity is set by the user and having something to read back out the set value doesn't seem too dangerous. It's also good for sanity checks and unit tests - e.g. as deserializing a stream summary gives back 0 capacity.

Reviewed By: Yuhta

Differential Revision: D68870501