refactor: use tryFindOrCompile instead of findOrCompile for regex [1/n] (#12234)
Summary:

When evaluating regexes and the regex throws, we can simply capture the exception inside a status instead of doing a throw which is extremely expensive. This change just uses the pre-existing expected API for error handling. This allows queries which times out after 2 hours on an operator to finish in 47 minutes.

Reviewed By: Yuhta

Differential Revision: D68983392