fix: Remove support for evaluating sub-optimial filters in joins
Summary:
This enables the hash join operator to throw an error when a filter is applied exclusively
to columns from one side of the join. This validation ensures the detection of
suboptimal plans, where filters are evaluated at the join operator instead of being
pushed to the upstream scan node of the respective join side for optimization.

Also adds a private API to ensure an internal state variable which is a buffer is
unique before any mutations are applied to it.

Differential Revision: D69085909