fix(fuzzer): Filter keys from both join sides instead of just one (#12246)
Summary:

Joins operators with filters on one side are not currently supported in Velox. It is expected that such a filter would be pushed down to the table scan and therefore not show up in the join operator.

The fix to support this would be disabling filter pushdown when filtering only one side of a join.

Alternatively, filters within join operators can be checked to ensure that the filter contains keys from both sides, else throw an exception. This change will require testing in production to see if Presto produces such plans.

For now, changing the join fuzzer filter creation to include keys from both sides will avoid this type of failure completely.

Differential Revision: D69079267