fix: Avoid redownload coalesced region gap twice in buffered inputs (#12018)
Summary:

In the current `DirectBufferedInput` and `CachedBufferedInput`, when we
coalesce some regions during prefetch, the gap between regions are not marked as
available, and if they are needed in adhoc streams, we would issue separate IOs
for them, resulting in unnecessary IOs.  This change fixes this by rearranging
the regions and loads generation logic, first generate and group the prefetch
regions, then we can compare the coalesced regions to the adhoc regions and move
those already coalesced to the prefetch loads, reducing the number of adhoc IOs.

Reviewed By: oerling

Differential Revision: D67810195