feat: Add DecodedVector::sharedBase() (#12249)
Summary:

Sometimes we need to take shared ownership of the base value vector of a dictionary.  The current `DecodedVector` only keeps reference to a raw pointer so there is no way to get hold of the `shared_ptr`.  We add `DecodedVector::sharedBase()` and overload of `DecodedVector::decode` to take `shared_ptr` so that we can get the shared ownership.

Reviewed By: bikramSingh91

Differential Revision: D69081492